victori/functions/requirements.txt
firebase_functions~=0.1.0
firebase-admin~=6.4.0
python-dateutil~=2.8.0
pydantic~=2.0.0
torch~=2.0.0
torchvision~=0.15.0
Pillow~=10.0.0